addresses some feedback
