+++ ui/App.svelte
@@ -12,7 +12,6 @@
   import CreateProject from "./Screen/CreateProject.svelte";
   import DesignSystemGuide from "./Screen/DesignSystemGuide.svelte";
   import Help from "./Screen/Help.svelte";
-  import CreateIdentitySuccess from "./Screen/IdentityCreationSuccess.svelte";
   import Network from "./Screen/Network.svelte";
   import NotFound from "./Screen/NotFound.svelte";
   import Project from "./Screen/Project.svelte";
@@ -41,7 +40,6 @@
 
   const routes = {
     "/identity/new": CreateIdentity,
-    "/identity/success": CreateIdentitySuccess,
     "/search": Search,
     "/network": Network,
     "/projects": Projects,
