+++ ui/Screen/DesignSystemGuide.svelte
@@ -529,6 +529,7 @@
         imageUrl="https://avatars.dicebear.com/v2/jdenticon/two.svg"
         variant="project"
         size="big" />
+      <Avatar size="huge" avatarFallback={avatarFallback2} />
     </Swatch>
 
     <Swatch>
