+++ ui/lib/hotkeys.js
@@ -18,12 +18,20 @@ export const initializeHotkeys = () => {
     push(path.help());
   });
 
+  hotkeys("shift+i", () => {
+    if (path.active(path.createIdentity(), get(location))) {
+      pop();
+    }
+    push(path.createIdentity());
+  });
+
   hotkeys("esc", () => {
     if (
       path.active(path.help(), get(location)) ||
       path.active(path.designSystemGuide(), get(location)) ||
       path.active(path.createProject(), get(location)) ||
-      path.active(path.registerProject("**"), get(location), true)
+      path.active(path.registerProject("**"), get(location), true) ||
+      path.active(path.createIdentity(), get(location))
     ) {
       pop();
     }
